Using system Crc32, removed ObfuscatedIP serializer
